% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{align_reads}
\alias{align_reads}
\title{Align Reads}
\usage{
align_reads(
  bwa_index_prefix,
  forward_reads,
  reverse_reads,
  outfile_prefix,
  bwa_options = "",
  threads = parallel::detectCores()
)
}
\arguments{
\item{bwa_index_prefix}{prefix of indexed reference genome (e.g. output of prep_for_alignment()) (string)}

\item{forward_reads}{path to fastq file containing forward reads (string)}

\item{reverse_reads}{path to fastq file containing reverse reads (string)}

\item{outfile_prefix}{prefix used for output alignment files}

\item{bwa_options}{supply any other options to bwa-mem2 mem command. For example '-k 50' will change min seed length to 50. (string)}

\item{threads}{how many cup threads to use (int)}
}
\value{
outfile path
}
\description{
Align Reads
}
\examples{
\dontrun{

prep_for_alignment()
}
}
