% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{bam_index}
\alias{bam_index}
\title{Bam Index}
\usage{
bam_index(path_to_sorted_bam, threads = parallel::detectCores())
}
\arguments{
\item{path_to_sorted_bam}{path to a coordinate sorted bam file (string)}

\item{threads}{number of cores to use for multithreading (int)}
}
\value{
path to sorted bam (string)
}
\description{
Index a bam file
}
\details{
Produce a bai index for a sorted bam. Requires samtools to be installed.
}
\examples{
\dontrun{
bam_index("path/to/sorted/bam")
}
}
