% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{choose_best_assembly}
\alias{choose_best_assembly}
\title{Best Assembly}
\usage{
choose_best_assembly(
  refseq_data_frame = load_refseq_data_frame_from_cache(),
  taxid_of_interest,
  intraspecific_filter = NA,
  break_ties_based_on_newest_sequence_added = TRUE,
  return_accession_only = TRUE
)
}
\arguments{
\item{refseq_data_frame}{from \strong{load_refseq_data_frame_from_cache()}}

\item{taxid_of_interest}{taxonomyID (int)}

\item{intraspecific_filter}{e.g. 'strain=Raji'. For options look at the infraspecific_name column of the refseq_data_frame (string)}

\item{break_ties_based_on_newest_sequence_added}{if we cant decide which is the best ref to use because they;re so similar - just pick the most recently uploaded sequence}
}
\value{

}
\description{
Best Assembly
}
