% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{delete_refseq_data_cache}
\alias{delete_refseq_data_cache}
\title{Delete the refseq Cache}
\usage{
delete_refseq_data_cache()
}
\description{
Remove existing cache. Good practice to run before uninstalling package.
}
