% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{download_assembly}
\alias{download_assembly}
\title{Download reference genome}
\usage{
download_assembly(
  target_assembly_accession,
  output_folder = getwd(),
  refseq_data_frame = load_refseq_data_frame_from_cache()
)
}
\arguments{
\item{target_assembly_accession}{accession of target assembly}

\item{output_folder}{directory to save assembly fasta file into (string)}

\item{refseq_data_frame}{from \strong{load_refseq_data_frame_from_cache()}}
}
\value{
filepath of downloaded file
}
\description{
Actually download the genome assembly
}
