% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{download_assembly_files}
\alias{download_assembly_files}
\title{download refseq assembly files}
\usage{
download_assembly_files(
  target_assembly_accession,
  assembly_file_type = c("genome", "stats", "genome_annotation"),
  output_folder = getwd(),
  refseq_data_frame = load_refseq_data_frame_from_cache()
)
}
\arguments{
\item{target_assembly_accession}{accession of assembly you're interested in. For example GCF_000371765.1 (string)}

\item{assembly_file_type}{type of file to download from refseq repo (string)}

\item{output_folder}{directory to output file into (string)}

\item{refseq_data_frame}{do not reccomend changing this. Automatically loads from \strong{load_refseq_data_frame_from_cache()}}
}
\value{
path to downloaded file
}
\description{
download refseq assembly files
}
\examples{
\dontrun{
download_assembly_files(
  target_assembly_accession = "GCF_000371765.1",
  assembly_file_type = "genome"
)
}

}
