% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{download_best_assembly}
\alias{download_best_assembly}
\title{Download_best_assembly}
\usage{
download_best_assembly(taxid_of_interest, ...)
}
\arguments{
\item{taxid_of_interest}{taxonomyID (int)}

\item{...}{
  Arguments passed on to \code{\link[=download_assembly]{download_assembly}}
  \describe{
    \item{\code{target_assembly_accession}}{accession of target assembly}
    \item{\code{output_folder}}{directory to save assembly fasta file into (string)}
    \item{\code{refseq_data_frame}}{from \strong{load_refseq_data_frame_from_cache()}}
  }}
}
\value{
filepath of downloaded file
}
\description{
Download_best_assembly
}
