% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{load_refseq_data_frame_from_cache}
\alias{load_refseq_data_frame_from_cache}
\title{Load data from cache}
\usage{
load_refseq_data_frame_from_cache()
}
\value{
data.frame
}
\description{
Loads a dataframe describing reference genomes and their ftp links
}
\examples{
refseq_data_frame = load_refseq_data_frame_from_cache()
}
