% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{load_refseq_data_frame_from_ftp}
\alias{load_refseq_data_frame_from_ftp}
\title{Cache refseq contents}
\usage{
load_refseq_data_frame_from_ftp()
}
\value{
refseq dataframe describing available assemblies
}
\description{
Download a file that lists metadata for all refseq entries (including ftp path).
We will use this later to download assemblies.
}
\examples{
\dontrun{
load_refseq_data_frame_from_ftp()
}
}
