% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{prep_for_alignment}
\alias{prep_for_alignment}
\title{Prep a refseq assembly for alignment}
\usage{
prep_for_alignment(path_to_assembly_fasta)
}
\arguments{
\item{path_to_assembly_fasta}{path to assembly fasta}
}
\value{
prefix of bwa-indexed (character)
}
\description{
Prep a refseq assembly for alignment
}
