% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{refseq_download_genome_assembly}
\alias{refseq_download_genome_assembly}
\title{Download reference genome}
\usage{
refseq_download_genome_assembly(
  target_assembly_accession,
  output_folder = getwd(),
  refseq_data_frame = load_refseq_data_frame_from_cache()
)
}
\arguments{
\item{target_assembly_accession}{accession of assembly you're interested in. For example GCF_000371765.1 (string)}

\item{output_folder}{directory to output file into (string)}

\item{refseq_data_frame}{do not reccomend changing this. Automatically loads from \strong{load_refseq_data_frame_from_cache()}}
}
\value{
path to downloaded file
}
\description{
Download the genome assembly
}
\examples{
\dontrun{
download_assembly_files(
  target_assembly_accession = "GCF_000371765.1",
  assembly_file_type = "genome"
)
}

}
