% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{update_refseq_data_cache}
\alias{update_refseq_data_cache}
\title{Update refseq data cache}
\usage{
update_refseq_data_cache()
}
\value{
NULL. Run for side-effects
}
\description{
Download a file that lists metadata for all refseq entries (including ftp path).
We will use this later to download assemblies.
}
\examples{
\dontrun{
update_refseq_data_cache()
}
}
