% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBestRefseqAssembly.R
\name{choose_best_assembly}
\alias{choose_best_assembly}
\title{Best Assembly}
\usage{
choose_best_assembly(
  taxid_of_interest,
  intraspecific_filter = NA,
  break_ties_based_on_newest_sequence_added = TRUE,
  return_accession_only = TRUE,
  refseq_data_frame = load_refseq_data_frame_from_cache()
)
}
\arguments{
\item{taxid_of_interest}{taxonomyID (int)}

\item{intraspecific_filter}{e.g. 'strain=Raji'. For options look at the infraspecific_name column of the refseq_data_frame (string)}

\item{break_ties_based_on_newest_sequence_added}{if we cant decide which is the best ref to use because they;re so similar - just pick the most recently uploaded sequence}

\item{return_accession_only}{Return just the best assembly accession. If false will return a dataframe with accession ID plus lots of other info (bool)}

\item{refseq_data_frame}{from \strong{load_refseq_data_frame_from_cache()}}
}
\value{
if return_accession_only=TRUE then best assembly accession (string). Otherwise returns dataframe with accession ID of best assembly plus lots of other info (bool)
}
\description{
Best Assembly
}
