% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{download_assembly}
\alias{download_assembly}
\title{Download reference genome}
\usage{
download_assembly(
  refseq_data_frame = load_refseq_data_frame_from_cache(),
  target_assembly_accession
)
}
\arguments{
\item{refseq_data_frame}{from \strong{load_refseq_data_frame_from_cache()}}

\item{target_assembly_accession}{accession of target assembly}

\item{outfile_dir}{directory to save assembly fasta file into}
}
\value{
Run for its side effects
}
\description{
Actually download the genome using aria2c
}
