% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRefseqGenomeHash.R
\name{write_refseq_dataframe}
\alias{write_refseq_dataframe}
\title{Write refseq dataframe}
\usage{
write_refseq_dataframe(refseq_data_frame)
}
\arguments{
\item{refseq_data_frame}{}
}
\description{
Write refseq dataframe
}
